// src/services/apis/payments/types/payment-error-enum.ts
export enum PaymentResponseCode {
  Success = "0",
  Unknown = "?",
  CardNotEnabled = "1",
  BankRejected = "2",
  NoBankResponse = "3",
  CardExpired = "4",
  InsufficientBalance = "5",
  BankProcessingError = "6",
  TransactionError = "7",
  InvalidCardNumber = "8",
  InvalidCardHolder = "9",
  CardLockedOrExpired = "10",
  CardNotRegistered = "11",
  InvalidExpiryDate = "12",
  ExceededLimit = "13",
  InsufficientFunds = "21",
  InvalidAccountInfo = "22",
  AccountLocked = "23",
  InvalidCardInfo = "24",
  InvalidOTP = "25",
  PaymentTimeout = "253",
  UserCancelled = "99",
  ThreeDSecureAuthFailed = "B",
  InvalidCSC = "E",
  ThreeDSecureVerificationFailed = "F",
  TransactionRejected = "Z",
}